Rework the way we handle events with Riemann
We now handle Results and Responses differently, passing them
through Riemann separately, instead of trying to do a lot of weird
stuff within the Riemann stream to do this. The behavior of this
stream stage may not be exactly what you might expect. See the
documentation string for beavis.rehabilitationsschrift/handle-event
for more information.